pyproject.toml
[project]
name = "GroqCLI-Chatbot"
version = "0.1.0"
description = "GroqCLI-Chatbot — A professional terminal-based AI chatbot using GroqCloud API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]

dependencies = [
    "groq>=0.12.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "pygments>=2.17.0",  # Syntax highlighting
    "prompt-toolkit>=3.0.43",  # Better input handling, keyboard shortcuts
]

[project.scripts]
groqcli-chatbot = "growcli.main:app"
groqcli = "growcli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ══════════════════════════════════════════════════════════
# THIS IS THE CRITICAL FIX — tells hatchling WHERE your
# source code lives, since project name (GroqCLI-Chatbot)
# does NOT match the package folder name (growcli)
# ══════════════════════════════════════════════════════════
[tool.hatch.build.targets.wheel]
packages = ["src/growcli"]

# ══════════════════════════════════════════════════════════
# FIXED: Use dependency-groups instead of deprecated
# tool.uv.dev-dependencies
# ══════════════════════════════════════════════════════════
[dependency-groups]
dev = [
    "pytest>=8.3.0",
]